+++ convex/mutations/organizations.ts
@@ -1,16 +1,18 @@
 // convex/organizations.ts
+import { getAuthUserId } from "@convex-dev/auth/server";
 import { internalMutation, mutation } from "../_generated/server";
 import { v } from "convex/values";
 
 export const createOrganization = internalMutation({
-  args: { 
-    name: v.string(), 
-    logo: v.string() },
+  args: {
+    name: v.string(),
+    logo: v.string(),
+  },
   handler: async (ctx, args) => {
-    const organizationId = await ctx.db.insert("organizations", { 
-        ...args, 
+    const organizationId = await ctx.db.insert("organizations", {
+      ...args,
     });
-    return organizationId
+    return organizationId;
   },
 });
 
@@ -23,19 +25,21 @@ export const updateOrganization = mutation({
     }),
   },
   handler: async (ctx, { organizationId, patch }) => {
+    const userId = await getAuthUserId(ctx);
+    if (!userId) throw new Error("Unauthorized");
+
     await ctx.db.patch(organizationId, patch);
     return { ok: true };
   },
 });
 
-
 export const deleteOrganization = mutation({
   args: { organizationId: v.id("organizations") },
   handler: async (ctx, { organizationId }) => {
+    const userId = await getAuthUserId(ctx);
+    if (!userId) throw new Error("Unauthorized");
+
     await ctx.db.delete(organizationId);
     return { ok: true };
   },
 });
-
-
-
